+++ ReferenceTests/src/tests/primitives.jl
@@ -1,6 +1,6 @@
 @reference_test "lines and linestyles" begin
     # For now disabled until we fix GLMakie linestyle
-    s = Scene(resolution = (800, 800), camera = campixel!)
+    s = Scene(size = (800, 800), camera = campixel!)
     scalar = 30
     points = Point2f[(1, 1), (1, 2), (2, 3), (2, 1)]
     linestyles = [
@@ -21,7 +21,7 @@
 end
 
 @reference_test "lines with gaps" begin
-    s = Scene(resolution = (800, 800), camera = campixel!)
+    s = Scene(size = (800, 800), camera = campixel!)
     points = [
         Point2f[(1, 0), (2, 0.5), (NaN, NaN), (4, 0.5), (5, 0)],
         Point2f[(NaN, NaN), (2, 0.5), (3, 0), (4, 0.5), (5, 0)],
@@ -37,7 +37,7 @@ end
 end
 
 @reference_test "scatters" begin
-    s = Scene(resolution = (800, 800), camera = campixel!)
+    s = Scene(size = (800, 800), camera = campixel!)
 
     markersizes = 0:2:30
     markers = [:circle, :rect, :cross, :utriangle, :dtriangle,
@@ -58,7 +58,7 @@ end
 end
 
 @reference_test "scatter rotations" begin
-    s = Scene(resolution = (800, 800), camera = campixel!)
+    s = Scene(size = (800, 800), camera = campixel!)
 
     rotations = range(0, 2pi, length = 15)
     markers = [:circle, :rect, :cross, :utriangle, :dtriangle,
@@ -82,7 +82,7 @@ end
 end
 
 @reference_test "scatter with stroke" begin
-    s = Scene(resolution = (350, 700), camera = campixel!)
+    s = Scene(size = (350, 700), camera = campixel!)
 
     # half stroke, half glow
     strokes = range(1, 4, length=7)
@@ -115,7 +115,7 @@ end
 end
 
 @reference_test "scatter with glow" begin
-    s = Scene(resolution = (350, 700), camera = campixel!)
+    s = Scene(size = (350, 700), camera = campixel!)
 
     # half stroke, half glow
     glows = range(4, 1, length=7)
@@ -151,7 +151,7 @@ end
 @reference_test "scatter image markers" begin
     pixel_types = [ RGBA, RGBAf, RGBA{Float16}, ARGB, ARGB{Float16}, RGB, RGBf, RGB{Float16} ]
     rotations = [ 2pi/3 * (i-1) for i = 1:length(pixel_types) ]
-    s = Scene(resolution = (100+100*length(pixel_types), 400), camera = campixel!)
+    s = Scene(size = (100+100*length(pixel_types), 400), camera = campixel!)
     filename = Makie.assetpath("icon_transparent.png")
     marker_image = load(filename)
     for (i, (rot, pxtype)) in enumerate(zip(rotations, pixel_types))
@@ -169,7 +169,7 @@ end
 
 
 @reference_test "basic polygon shapes" begin
-    s = Scene(resolution = (800, 800), camera = campixel!)
+    s = Scene(size = (800, 800), camera = campixel!)
     scalefactor = 70
     Pol = Makie.GeometryBasics.Polygon
     polys = [
@@ -222,7 +222,7 @@ end
 
 
 @reference_test "BezierPath markers" begin
-    f = Figure(resolution = (800, 800))
+    f = Figure(size = (800, 800))
     ax = Axis(f[1, 1])
 
     markers = [
@@ -268,7 +268,7 @@ end
 end
 
 @reference_test "BezierPath marker stroke" begin
-    f = Figure(resolution = (800, 800))
+    f = Figure(size = (800, 800))
     ax = Axis(f[1, 1])
 
     # Same as above
@@ -290,7 +290,7 @@ end
 
 
 @reference_test "complex_bezier_markers" begin
-    f = Figure(resolution = (800, 800))
+    f = Figure(size = (800, 800))
     ax = Axis(f[1, 1])
 
     arrow = BezierPath([
@@ -425,7 +425,7 @@ function draw_marker_test!(scene, marker, center; markersize=300)
 end
 
 @reference_test "marke glyph alignment" begin
-    scene = Scene(resolution=(1200, 1200))
+    scene = Scene(size=(1200, 1200))
     campixel!(scene)
     # marker is in front, so it should not be smaller than the background rectangle
     plot_row!(scene, 0, false)
@@ -464,7 +464,7 @@ end
 end
 
 @reference_test "barplot with TeX-ed labels" begin
-    fig = Figure(resolution = (800, 800))
+    fig = Figure(size = (800, 800))
     lab1 = L"\int f(x) dx"
     lab2 = lab1
     # lab2 = L"\frac{a}{b} - \sqrt{b}" # this will not work until #2667 is fixed